:bug: Fix price attributes when next-day prices are present
Generate different sets of sensor attributes for hourly prices for current day
and for the next day (available at evening), so attrs like `price_position`
or `price_ratio` don't change for the current day when next-day prices are received.
