+++ tests/test_tariff_attributes.py
@@ -5,7 +5,7 @@ import holidays
 import pytest
 
 from aiopvpc.const import REFERENCE_TZ
-from aiopvpc.pvpc_data import _tariff_period_key
+from aiopvpc.pvpc_tariff import get_current_and_next_tariff_periods
 
 _HOLIDAYS_VERSION = tuple(map(int, holidays.__version__.split(".")))
 
@@ -23,7 +23,8 @@ def test_number_of_national_holidays(year, days_weekend_p3, extra_days_p3):
     holidays_p3 = weekend_days_p3 = 0
     day = datetime(year, 1, 1, 15, tzinfo=REFERENCE_TZ)
     while day.year == year:
-        if _tariff_period_key(day, False) == "P3":
+        period, _, _ = get_current_and_next_tariff_periods(day, False)
+        if period == "P3":
             if day.isoweekday() > 5:
                 weekend_days_p3 += 1
             else:
